Add Stuctured Data and update Next.js
* Add Structured Data Testing Tool

Now displays structured data found.

* Update to Next.js 9

This update means local development is much faster, as it starts up  in dev mode much quicker.

It also simplifies the file structure and reduces the configuration required for the app.

However, `next start` it works slightly differently from `now dev` and this has introduced some issues with dependancies.

* Remove WHOIS results

This is (hopefully) temporary due to issue that came up after upgrading to Next.js 9 and webpack not knowing how to handle shell scripts out of the box.

* Updated unfluff fork

Minor update to unfluff fork to work with Next.js 9 (changed readFile to readFileSync).